Extract broser functionality to their own files
