+++ filter.c
@@ -491,103 +491,6 @@ void filterPost( register int inputChar )
    }
 }
 
-void filterUrl( const char *aryLine )
-{
-   static int multiline = 0;
-   static char aryUrlBuffer[1024];
-   char *ptrCursor, *ptrNext;
-
-   if ( !urlQueue )
-   { /* Can't store URLs for some reason */
-      return;
-   }
-
-   if ( !multiline )
-   {
-      snprintf( aryUrlBuffer, sizeof( aryUrlBuffer ), "%s", aryLine );
-   }
-   else
-   {
-      size_t ulen = strlen( aryUrlBuffer );
-      if ( ulen < sizeof( aryUrlBuffer ) - 1 )
-      {
-         snprintf( aryUrlBuffer + ulen, sizeof( aryUrlBuffer ) - ulen, "%s", aryLine );
-      }
-   }
-
-   {
-      size_t urlLength = strlen( aryUrlBuffer );
-      while ( urlLength > 0 )
-      {
-         size_t index = urlLength - 1;
-         if ( aryUrlBuffer[index] == ' ' || aryUrlBuffer[index] == '\t' || aryUrlBuffer[index] == '\r' )
-         {
-            aryUrlBuffer[index] = 0;
-         }
-         else
-         {
-            break;
-         }
-         urlLength = index;
-      }
-   }
-   {
-      size_t ulen = strlen( aryUrlBuffer );
-      if ( ulen < sizeof( aryUrlBuffer ) - 1 )
-      {
-         snprintf( aryUrlBuffer + ulen, sizeof( aryUrlBuffer ) - ulen, " " );
-      }
-   }
-
-   if ( !( ptrCursor = strstr( aryUrlBuffer, "http://" ) ) )
-   {
-      if ( !( ptrCursor = strstr( aryUrlBuffer, "ftp://" ) ) )
-      {
-         aryUrlBuffer[0] = 0;
-         multiline = 0;
-         return;
-      }
-   }
-   /* offset unused */
-
-   for ( ptrNext = ptrCursor; *ptrNext; ptrNext++ )
-   {
-      if ( findChar( ":/abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789$-_@.&+,=;?%~{|}", *ptrNext ) )
-      {
-         continue;
-      }
-      *ptrNext = 0;
-      /* length unused */
-
-      /* Oops, looks like a multi-aryLine URL */
-      if ( ( !multiline && ptrCursor == aryLine &&
-             ptrNext > ptrCursor + 77 ) ||
-           ( multiline && strlen( aryLine ) > 77 ) )
-      {
-         if ( strlen( aryLine ) > 77 )
-         {
-            break;
-         }
-      }
-
-      if ( !isQueued( ptrCursor, urlQueue ) )
-      {
-         char aryTempText[1024];
-         while ( !pushQueue( ptrCursor, urlQueue ) )
-         {
-            popQueue( aryTempText, urlQueue );
-         }
-      }
-      /*	    printf("\r\nSnarfed URL: <%s>\r\n", urls[nurls - 1]); */
-      multiline = 0;
-      return;
-   }
-
-   multiline = 1;
-   /*	printf("Multiline URL, got <%s> so far.\r\n", aryUrlBuffer); */
-   return;
-}
-
 void filterData( register int inputChar )
 {
    static int ansistate = 0; /* Counter for ANSI transformations */
